debug cuda dfs

+++ iGlobOptCudaDFS/include/CUDAGlobalOptimization.h
@@ -533,7 +533,7 @@ void fnGetOptValueWithCUDA_deep(double *inBox, int inRank, int inNumBoxesSplitCo
 }
 
 
-__const__ double rank = 10;
+__const__ int rank = 10;
 
 
 __global__ void globOptCUDA(double *inBox, int inRank, int *workLen, double *min, double inRec, double inEps)
